startup-operator/requirements.txt
openai>=1.40.0
notion-client>=2.2.1
python-dotenv>=1.0.1
